requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.4
fastapi==0.99.1
h11==0.14.0
idna==3.4
pydantic==1.10.11
pydantic_core==2.1.2
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
